Refactor more detail view
