+++ frontend/src/components/OfferedServiceCard/MoreDetailView.js
@@ -1,87 +1,9 @@
 import React, { Component } from 'react';
 import Collapse from '@material-ui/core/Collapse/Collapse';
 import Slide from '@material-ui/core/Slide/Slide';
-import FormControlLabel from '@material-ui/core/FormControlLabel/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox/Checkbox';
-import { green, grey } from '@material-ui/core/colors';
-import { withStyles } from '@material-ui/core';
 import Images from '../../assets/Images';
 import BottomButton from '../BottomButton';
-
-const styles = {
-  root: {
-    '&$checked': {
-      color: green[500]
-    }
-  },
-  checked: {
-    color: green[500]
-  },
-  disabled: {
-    '&$checked': {
-      color: grey[500]
-    }
-  }
-};
-
-const CheckBoxes = ({
-  termsChecked,
-  consentChecked,
-  handleChange,
-  disabled,
-  classes
-}) => (
-  <div className="offered-service-card__row--checkboxes">
-    <FormControlLabel
-      control={
-        <Checkbox
-          checked={termsChecked}
-          onChange={() => handleChange('termsChecked')}
-          classes={{
-            root: classes.root,
-            checked: classes.checked,
-            disabled: classes.disabled
-          }}
-          disabled={disabled}
-        />
-      }
-      label={
-        <div>
-          <span className="body-text">Please read our </span>
-          <a href="/" className="body-text">
-            <b>terms and conditions</b>
-          </a>
-          <span className="body-text"> and view our </span>
-          <a href="/" className="body-text">
-            <b>privacy policy here *</b>
-          </a>
-        </div>
-      }
-    />
-    <FormControlLabel
-      control={
-        <Checkbox
-          checked={consentChecked}
-          onChange={() => handleChange('consentChecked')}
-          classes={{
-            root: classes.root,
-            checked: classes.checked,
-            disabled: classes.disabled
-          }}
-          disabled={disabled}
-        />
-      }
-      label={
-        <div>
-          <span className="body-text">
-            I have read and understood the terms and conditions and consent to
-            my details being stored *
-          </span>
-        </div>
-      }
-    />
-  </div>
-);
+import CheckboxesSection from './CheckboxesSection';
 
 const DetailItem = ({ title, description }) => (
   <div className="offered-service-card__row offered-service-card__row--detail">
@@ -164,7 +86,7 @@ class MoreDetailView extends Component {
                 description={requiredSensors}
               />
 
-              <CheckBoxes
+              <CheckboxesSection
                 termsChecked={termsChecked}
                 consentChecked={consentChecked}
                 handleChange={this.handleCheckChange}
@@ -187,4 +109,4 @@ class MoreDetailView extends Component {
   }
 }
 
-export default withStyles(styles)(MoreDetailView);
+export default MoreDetailView;
